feat: added filter unit tests
