+++ online-store/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "parser": "@typescript-eslint/parser, ",
+  "parser": "@typescript-eslint/parser",
   "plugins": ["prettier", "import", "@typescript-eslint"],
   "extends": [
       "eslint:recommended",
