+++ online-store/src/components/util/filterLogic/filterBrand.ts
@@ -1,0 +1,10 @@
+import { BrandType, Product } from '../../../domain/model';
+
+export const filterByBrand = (brands: BrandType[], data: Product[]): Product[] => {
+  //console.log(brands);
+  if (brands.length === 0) {
+    return data;
+  }
+  const filteredData = data.filter((product) => brands.includes(product.brand));
+  return filteredData;
+};
