--- online-store/tests/sum.test.ts
@@ -1,8 +1,0 @@
-/* eslint-disable no-undef */
-
-import { isInt } from './sum';
-
-test('name', () => {
-  expect(1).toBe(1);
-  expect(isInt()).toBe(13);
-});
